mi-proyecto-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
tensorflow==2.16.2
numpy==1.24.3
python-multipart==0.0.6
aiohttp==3.9.1
openai==1.3.0
httpx==0.27.2
google-generativeai==0.3.0
requests==2.31.0
# Data processing and export
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9
scikit-learn==1.3.2
